finch_bot/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0